+++ libcrux-ml-dsa/cg/libcrux_mldsa65_avx2.h
@@ -6456,7 +6456,7 @@ libcrux_ml_dsa_encoding_signature_deserialize_64(
         hint_serialized, max_ones_in_hint + i1, uint8_t);
     if (current_true_hints_seen < previous_true_hints_seen) {
       malformed_hint = true;
-    } else if (previous_true_hints_seen > max_ones_in_hint) {
+    } else if (current_true_hints_seen > max_ones_in_hint) {
       malformed_hint = true;
     } else {
       for (size_t i = previous_true_hints_seen; i < current_true_hints_seen;