+++ libcrux-ml-dsa/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- [#1348](https://github.com/cryspen/libcrux/pull/1348): Fix incorrect hint decoding
 - [#1347](https://github.com/cryspen/libcrux/pull/1347): Fix incorrect norm check of signer response during verification
 
 ## [0.0.7] (2026-02-12)